LayoutTests/media/wireless-playback-media-player/load-url-src-attr-expected.txt

Test that a video's src attribute URL is passed to a MediaDeviceRoute when the route activates.


EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
RUN(internals.mockMediaDeviceRouteController.activateRoute(route))
EXPECTED (actualURL == '/LayoutTests/media/content/test.mp4') OK
END OF TEST
